refactor(audit): extract finding prioritization into domain layer
Move pure prioritization business rules out of the presentation/builder layer
into a new `audit::prioritization` domain module:

- Enums Priority, Effort, ExecutionPriority (with their en/de label methods)
  move from output/report_model.rs; report_model re-exports them so existing
  `report_model::{Priority, Effort, ExecutionPriority}` paths keep working.
- Functions severity_to_priority, score_to_priority, derive_execution_priority
  move from output/builder/actions.rs; batch.rs and single/findings.rs now
  import them from audit::prioritization.

These are locale-neutral rules over domain types (Severity, score, dimension),
so the domain is now their home and the builder is a pure mapper for them. No
behaviour change — output is byte-identical; adds unit tests for the rules.
Dependency direction stays correct (output -> audit, never the reverse).
